backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.29.0
certifi==2026.1.4
click==8.3.1
fastapi==0.109.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
PyYAML==6.0.3
redis==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.15.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
